# Additional encoding/decoding optimization