… decode.go
… encode.go
… types.go