* Replace reader/writer interfaces with concrete `buffer` type.
* Avoid intermediate buffer during map encoding.
* Avoid `milliseconds` marshal allocation.
* Remove `errNull`/`isNull` from marshal and avoid defer in hot path.
* Cleanup of decoding code.
* Minimize bounds checks by adding hints.
* Check interface types last in type switches.
* Add fastpath in unmarshal for `**uint32`.

Closes #17 

```
name                                       old time/op    new time/op    delta
FrameMarshal/transfer-8                       420ns ± 4%     197ns ± 5%   -53.04%  (p=0.000 n=10+10)
FrameUnmarshal/transfer-8                    1.84µs ± 0%    0.61µs ± 5%   -66.64%  (p=0.000 n=8+10)
Marshal/*amqp.performOpen-8                   787ns ± 3%     293ns ± 5%   -62.83%  (p=0.000 n=10+10)
Marshal/*amqp.performBegin-8                  621ns ± 1%     216ns ± 3%   -65.27%  (p=0.000 n=10+9)
Marshal/*amqp.performAttach-8                2.45µs ± 1%    1.12µs ± 3%   -54.13%  (p=0.000 n=10+10)
Marshal/amqp.role-8                          30.5ns ± 2%    20.5ns ± 2%   -32.75%  (p=0.000 n=10+10)
Marshal/*amqp.unsettled-8                     258ns ± 8%     100ns ± 3%   -61.10%  (p=0.000 n=10+9)
Marshal/*amqp.source-8                        900ns ± 1%     401ns ± 5%   -55.43%  (p=0.000 n=10+10)
Marshal/*amqp.target-8                        516ns ± 1%     244ns ± 3%   -52.79%  (p=0.000 n=10+10)
Marshal/*amqp.performFlow-8                   576ns ± 1%     193ns ± 4%   -66.42%  (p=0.000 n=10+10)
Marshal/*amqp.performTransfer-8               404ns ± 6%     184ns ± 2%   -54.35%  (p=0.000 n=10+10)
Marshal/*amqp.performDisposition-8            215ns ± 5%     102ns ± 4%   -52.54%  (p=0.000 n=10+10)
Marshal/*amqp.performDetach-8                 612ns ± 2%     279ns ± 0%   -54.37%  (p=0.000 n=10+6)
Marshal/*amqp.performDetach#01-8             1.23µs ± 8%    0.42µs ± 4%   -65.71%  (p=0.000 n=9+10)
Marshal/amqp.ErrorCondition-8                69.9ns ± 1%    27.3ns ± 4%   -60.96%  (p=0.000 n=10+10)
Marshal/*amqp.Error-8                         493ns ± 1%     221ns ± 5%   -55.12%  (p=0.000 n=10+10)
Marshal/*amqp.performEnd-8                    595ns ± 1%     264ns ± 3%   -55.69%  (p=0.000 n=8+10)
Marshal/*amqp.performClose-8                  601ns ± 2%     263ns ± 4%   -56.29%  (p=0.000 n=10+10)
Marshal/*amqp.Message-8                      1.83µs ± 0%    0.80µs ± 4%   -56.33%  (p=0.000 n=8+10)
Marshal/*amqp.MessageHeader-8                 194ns ± 2%      82ns ± 3%   -57.39%  (p=0.000 n=10+10)
Marshal/*amqp.MessageProperties-8             546ns ± 3%     264ns ± 5%   -51.56%  (p=0.000 n=8+10)
Marshal/*amqp.stateReceived-8                 137ns ± 3%      48ns ± 2%   -65.10%  (p=0.000 n=9+10)
Marshal/*amqp.stateAccepted-8                30.1ns ± 2%    21.4ns ± 4%   -28.80%  (p=0.000 n=8+10)
Marshal/*amqp.stateRejected-8                 608ns ± 1%     261ns ± 3%   -57.04%  (p=0.000 n=10+9)
Marshal/*amqp.stateReleased-8                30.6ns ± 5%    22.2ns ± 3%   -27.30%  (p=0.000 n=8+10)
Marshal/*amqp.stateModified-8                 352ns ± 4%     145ns ± 6%   -58.84%  (p=0.000 n=10+10)
Marshal/amqp.lifetimePolicy-8                29.9ns ± 3%    19.8ns ± 3%   -33.73%  (p=0.000 n=10+10)
Marshal/amqp.SenderSettleMode-8              35.1ns ± 3%    23.7ns ± 3%   -32.44%  (p=0.000 n=9+9)
Marshal/amqp.ReceiverSettleMode-8            36.3ns ±13%    23.6ns ± 4%   -35.07%  (p=0.000 n=10+10)
Marshal/*amqp.saslInit-8                      160ns ± 5%      90ns ± 4%   -43.69%  (p=0.000 n=8+10)
Marshal/*amqp.saslMechanisms-8                235ns ± 1%      74ns ± 7%   -68.69%  (p=0.000 n=10+10)
Marshal/*amqp.saslOutcome-8                   169ns ± 1%      71ns ± 9%   -57.88%  (p=0.000 n=10+10)
Marshal/amqp.milliseconds-8                  60.2ns ± 2%    19.0ns ± 5%   -68.40%  (p=0.000 n=10+10)
Marshal/amqp.symbol-8                        29.0ns ± 5%    25.1ns ± 5%   -13.40%  (p=0.000 n=10+10)
Marshal/map[amqp.symbol]interface_{}-8        242ns ± 9%      81ns ± 2%   -66.57%  (p=0.000 n=10+9)
Marshal/amqp.UUID-8                          46.4ns ± 2%    19.2ns ± 6%   -58.73%  (p=0.000 n=9+10)
Marshal/bool-8                               16.6ns ± 5%     6.2ns ± 4%   -62.82%  (p=0.000 n=10+10)
Marshal/int8-8                               19.4ns ± 1%     8.3ns ± 5%   -57.28%  (p=0.000 n=9+10)
Marshal/int8#01-8                            19.5ns ± 2%     8.4ns ± 4%   -57.10%  (p=0.000 n=9+10)
Marshal/int16-8                              24.3ns ± 6%     6.8ns ± 5%   -72.13%  (p=0.000 n=10+10)
Marshal/int16#01-8                           24.9ns ± 2%     7.0ns ± 6%   -71.90%  (p=0.000 n=9+10)
Marshal/int32-8                              33.1ns ± 2%     6.9ns ± 3%   -79.13%  (p=0.000 n=9+10)
Marshal/int32#01-8                           32.6ns ± 5%     6.9ns ± 5%   -78.94%  (p=0.000 n=10+10)
Marshal/int64-8                              45.8ns ± 4%     7.0ns ± 4%   -84.66%  (p=0.000 n=10+10)
Marshal/int64#01-8                           45.7ns ± 3%     7.1ns ± 5%   -84.54%  (p=0.000 n=8+10)
Marshal/uint8-8                              19.0ns ± 2%     8.3ns ± 5%   -56.49%  (p=0.000 n=9+10)
Marshal/uint16-8                             24.8ns ± 4%     7.4ns ± 2%   -70.23%  (p=0.000 n=10+9)
Marshal/uint32-8                             33.4ns ± 3%     7.1ns ± 5%   -78.78%  (p=0.000 n=8+10)
Marshal/uint64-8                             46.5ns ± 3%     8.5ns ± 4%   -81.78%  (p=0.000 n=9+10)
Marshal/float32-8                            33.7ns ± 7%     7.2ns ± 6%   -78.73%  (p=0.000 n=10+10)
Marshal/float32#01-8                         33.5ns ± 3%     7.1ns ± 5%   -78.89%  (p=0.000 n=10+10)
Marshal/float32#02-8                         33.4ns ± 5%     7.0ns ± 5%   -78.99%  (p=0.000 n=10+10)
Marshal/float32#03-8                         34.1ns ± 6%     7.2ns ± 4%   -78.86%  (p=0.000 n=10+10)
Marshal/float64-8                            47.3ns ± 3%     7.4ns ± 4%   -84.38%  (p=0.000 n=9+10)
Marshal/float64#01-8                         47.3ns ± 3%     7.4ns ± 6%   -84.37%  (p=0.000 n=8+10)
Marshal/float64#02-8                         46.9ns ± 5%     7.6ns ± 1%   -83.87%  (p=0.000 n=10+10)
Marshal/float64#03-8                         47.6ns ± 4%     7.3ns ± 4%   -84.57%  (p=0.000 n=9+10)
Marshal/amqp.describedType-8                  119ns ± 5%      65ns ± 5%   -45.63%  (p=0.000 n=10+9)
Marshal/map[interface_{}]interface_{}-8       253ns ± 9%      79ns ± 4%   -68.82%  (p=0.000 n=9+10)
Marshal/map[string]interface_{}-8             252ns ±11%      83ns ± 3%   -67.08%  (p=0.000 n=10+9)
Marshal/amqp.ArrayUByte-8                    38.5ns ± 4%    29.6ns ± 1%   -23.25%  (p=0.000 n=10+9)
Marshal/[]int8-8                             47.2ns ± 5%    24.9ns ± 5%   -47.28%  (p=0.000 n=10+9)
Marshal/[]uint16-8                           75.1ns ± 3%    24.1ns ± 5%   -67.84%  (p=0.000 n=10+10)
Marshal/[]uint16#01-8                        75.4ns ± 4%    24.4ns ± 5%   -67.60%  (p=0.000 n=10+10)
Marshal/[]int16-8                            75.0ns ± 6%    24.6ns ± 4%   -67.26%  (p=0.000 n=10+10)
Marshal/[]int16#01-8                         73.7ns ± 4%    23.9ns ± 7%   -67.64%  (p=0.000 n=10+10)
Marshal/[]uint32-8                            106ns ± 4%      24ns ± 2%   -77.50%  (p=0.000 n=10+9)
Marshal/[]uint32#01-8                        49.6ns ± 5%    26.7ns ± 6%   -46.18%  (p=0.000 n=10+10)
Marshal/[]int32-8                             105ns ± 3%      23ns ± 4%   -77.70%  (p=0.000 n=10+8)
Marshal/[]int32#01-8                         49.0ns ± 3%    27.1ns ± 5%   -44.69%  (p=0.000 n=9+9)
Marshal/[]uint64-8                            172ns ± 4%      24ns ± 3%   -85.81%  (p=0.000 n=10+10)
Marshal/[]uint64#01-8                        49.2ns ± 3%    26.7ns ± 2%   -45.72%  (p=0.000 n=10+8)
Marshal/[]int64-8                             172ns ± 2%      25ns ± 3%   -85.57%  (p=0.000 n=8+9)
Marshal/[]int64#01-8                         49.1ns ± 4%    27.2ns ± 4%   -44.62%  (p=0.000 n=10+10)
Marshal/[]float32-8                          90.6ns ± 3%    23.0ns ± 3%   -74.58%  (p=0.000 n=9+9)
Marshal/[]float64-8                           143ns ± 6%      23ns ± 9%   -83.87%  (p=0.000 n=10+10)
Marshal/[]bool-8                             44.0ns ± 4%    21.8ns ±10%   -50.45%  (p=0.000 n=10+10)
Marshal/[]string-8                            131ns ± 6%      43ns ± 3%   -67.25%  (p=0.000 n=10+10)
Marshal/[]amqp.symbol-8                       130ns ± 3%      41ns ± 2%   -68.39%  (p=0.000 n=10+10)
Marshal/[][]uint8-8                           127ns ± 4%      37ns ± 6%   -71.02%  (p=0.000 n=10+8)
Marshal/[]time.Time-8                        58.9ns ± 4%    19.5ns ± 4%   -66.87%  (p=0.000 n=10+10)
Marshal/[]amqp.UUID-8                        55.0ns ± 2%    20.2ns ± 5%   -63.20%  (p=0.000 n=10+10)
Marshal/[]interface_{}-8                      125ns ± 3%      43ns ± 5%   -65.89%  (p=0.000 n=8+10)
Unmarshal/*amqp.performOpen-8                1.65µs ± 3%    0.91µs ±15%   -44.61%  (p=0.000 n=10+10)
Unmarshal/*amqp.performBegin-8               1.33µs ± 4%    0.78µs ±14%   -41.35%  (p=0.000 n=10+10)
Unmarshal/*amqp.performAttach-8              6.09µs ± 6%    3.80µs ±17%   -37.52%  (p=0.000 n=10+10)
Unmarshal/amqp.role-8                         215ns ±14%      58ns ± 8%   -73.18%  (p=0.000 n=10+10)
Unmarshal/*amqp.unsettled-8                   938ns ±20%     629ns ± 9%   -33.01%  (p=0.000 n=10+9)
Unmarshal/*amqp.source-8                     2.19µs ±11%    1.43µs ±25%   -34.87%  (p=0.000 n=10+10)
Unmarshal/*amqp.target-8                     1.29µs ± 5%    0.78µs ± 9%   -39.50%  (p=0.000 n=10+10)
Unmarshal/*amqp.performFlow-8                1.82µs ± 1%    0.77µs ±15%   -57.89%  (p=0.000 n=10+10)
Unmarshal/*amqp.performTransfer-8            1.71µs ± 2%    0.47µs ± 1%   -72.26%  (p=0.000 n=10+10)
Unmarshal/*amqp.performDisposition-8          910ns ± 1%     242ns ± 1%   -73.37%  (p=0.000 n=10+10)
Unmarshal/*amqp.performDetach-8              1.49µs ± 3%    0.85µs ± 9%   -42.86%  (p=0.000 n=10+9)
Unmarshal/*amqp.performDetach#01-8           1.93µs ± 7%    1.22µs ±13%   -36.82%  (p=0.000 n=10+10)
Unmarshal/amqp.ErrorCondition-8               271ns ±15%      92ns ±14%   -65.93%  (p=0.000 n=10+10)
Unmarshal/*amqp.Error-8                      1.19µs ± 9%    0.76µs ±17%   -36.36%  (p=0.000 n=10+10)
Unmarshal/*amqp.performEnd-8                 1.38µs ± 4%    0.86µs ±23%   -37.86%  (p=0.000 n=10+10)
Unmarshal/*amqp.performClose-8               1.32µs ± 2%    0.82µs ± 8%   -38.12%  (p=0.000 n=8+10)
Unmarshal/*amqp.Message-8                    4.66µs ± 4%    2.94µs ±16%   -36.87%  (p=0.000 n=10+10)
Unmarshal/*amqp.MessageHeader-8               647ns ± 0%     206ns ± 1%   -68.18%  (p=0.000 n=8+9)
Unmarshal/*amqp.MessageProperties-8          1.47µs ± 1%    0.53µs ± 1%   -63.75%  (p=0.000 n=10+10)
Unmarshal/*amqp.stateReceived-8               385ns ± 2%     126ns ± 1%   -67.23%  (p=0.000 n=9+10)
Unmarshal/*amqp.stateAccepted-8               255ns ±17%      94ns ± 3%   -63.08%  (p=0.000 n=10+10)
Unmarshal/*amqp.stateRejected-8              1.40µs ± 4%    0.87µs ±11%   -38.12%  (p=0.000 n=9+10)
Unmarshal/*amqp.stateReleased-8               251ns ± 9%      94ns ± 4%   -62.81%  (p=0.000 n=10+10)
Unmarshal/*amqp.stateModified-8              1.01µs ± 8%    0.68µs ±21%   -33.07%  (p=0.000 n=9+10)
Unmarshal/amqp.lifetimePolicy-8               203ns ± 7%      63ns ± 7%   -68.74%  (p=0.000 n=10+9)
Unmarshal/amqp.SenderSettleMode-8             218ns ± 9%      57ns ±21%   -73.84%  (p=0.000 n=9+10)
Unmarshal/amqp.ReceiverSettleMode-8           221ns ±11%      57ns ± 7%   -74.38%  (p=0.000 n=10+9)
Unmarshal/*amqp.saslInit-8                    551ns ± 1%     220ns ± 1%   -60.10%  (p=0.000 n=10+10)
Unmarshal/*amqp.saslMechanisms-8              420ns ± 2%     187ns ± 1%   -55.35%  (p=0.000 n=10+9)
Unmarshal/*amqp.saslOutcome-8                 511ns ± 2%     187ns ± 2%   -63.46%  (p=0.000 n=10+9)
Unmarshal/amqp.milliseconds-8                 189ns ±22%      60ns ±17%   -68.31%  (p=0.000 n=10+10)
Unmarshal/amqp.symbol-8                       201ns ±12%      74ns ±21%   -63.44%  (p=0.000 n=10+10)
Unmarshal/map[amqp.symbol]interface_{}-8      897ns ±31%     614ns ±16%   -31.57%  (p=0.000 n=10+10)
Unmarshal/amqp.UUID-8                         215ns ±16%      66ns ±12%   -69.16%  (p=0.000 n=10+10)
Unmarshal/bool-8                              174ns ±34%      56ns ±10%   -67.75%  (p=0.000 n=10+10)
Unmarshal/int8-8                              201ns ±26%      54ns ±17%   -72.95%  (p=0.000 n=10+10)
Unmarshal/int8#01-8                           175ns ±23%      57ns ±21%   -67.47%  (p=0.000 n=10+10)
Unmarshal/int16-8                             189ns ±15%      56ns ±16%   -70.29%  (p=0.000 n=10+10)
Unmarshal/int16#01-8                          185ns ±22%      55ns ±17%   -70.31%  (p=0.000 n=10+10)
Unmarshal/int32-8                             198ns ±13%      53ns ±11%   -73.27%  (p=0.000 n=10+9)
Unmarshal/int32#01-8                          200ns ±14%      55ns ±13%   -72.32%  (p=0.000 n=10+10)
Unmarshal/int64-8                             206ns ±16%      56ns ±19%   -72.93%  (p=0.000 n=10+10)
Unmarshal/int64#01-8                          190ns ±14%      52ns ±12%   -72.48%  (p=0.000 n=10+9)
Unmarshal/uint8-8                             188ns ±11%      55ns ±13%   -70.70%  (p=0.000 n=10+10)
Unmarshal/uint16-8                            197ns ±12%      55ns ±11%   -72.16%  (p=0.000 n=10+10)
Unmarshal/uint32-8                            187ns ±14%      57ns ±10%   -69.62%  (p=0.000 n=8+9)
Unmarshal/uint64-8                            174ns ±23%      54ns ±27%   -68.90%  (p=0.000 n=10+10)
Unmarshal/float32-8                           198ns ±24%      55ns ±20%   -72.17%  (p=0.000 n=10+10)
Unmarshal/float32#01-8                        187ns ±15%      55ns ±12%   -70.82%  (p=0.000 n=10+10)
Unmarshal/float32#02-8                        186ns ±19%      56ns ±13%   -69.76%  (p=0.000 n=10+10)
Unmarshal/float32#03-8                        184ns ±12%      52ns ±13%   -71.93%  (p=0.000 n=10+10)
Unmarshal/float64-8                           179ns ± 7%      54ns ± 7%   -69.55%  (p=0.000 n=9+10)
Unmarshal/float64#01-8                        188ns ±17%      54ns ±16%   -71.39%  (p=0.000 n=10+10)
Unmarshal/float64#02-8                        186ns ±15%      55ns ±19%   -70.35%  (p=0.000 n=10+10)
Unmarshal/float64#03-8                        185ns ±13%      56ns ±21%   -69.69%  (p=0.000 n=10+10)
Unmarshal/amqp.describedType-8                413ns ±12%     203ns ±11%   -50.91%  (p=0.000 n=10+10)
Unmarshal/map[interface_{}]interface_{}-8     774ns ±22%     618ns ±17%   -20.15%  (p=0.000 n=10+10)
Unmarshal/map[string]interface_{}-8           769ns ±35%     591ns ±17%   -23.16%  (p=0.015 n=10+10)
Unmarshal/amqp.ArrayUByte-8                   203ns ±16%      88ns ±10%   -56.70%  (p=0.000 n=10+10)
Unmarshal/[]int8-8                            205ns ±12%      58ns ±10%   -71.96%  (p=0.000 n=9+9)
Unmarshal/[]uint16-8                          194ns ±14%      63ns ±18%   -67.66%  (p=0.000 n=10+10)
Unmarshal/[]uint16#01-8                       194ns ±20%      62ns ±12%   -68.16%  (p=0.000 n=10+10)
Unmarshal/[]int16-8                           200ns ±12%      61ns ± 6%   -69.48%  (p=0.000 n=10+9)
Unmarshal/[]int16#01-8                        200ns ± 7%      64ns ±14%   -68.04%  (p=0.000 n=10+10)
Unmarshal/[]uint32-8                          193ns ±16%      61ns ±10%   -68.53%  (p=0.000 n=10+10)
Unmarshal/[]uint32#01-8                       200ns ±20%      58ns ±11%   -70.86%  (p=0.000 n=10+10)
Unmarshal/[]int32-8                           201ns ±20%      64ns ±15%   -68.14%  (p=0.000 n=10+10)
Unmarshal/[]int32#01-8                        187ns ±14%      60ns ±17%   -67.67%  (p=0.000 n=9+10)
Unmarshal/[]uint64-8                          182ns ±12%      63ns ±12%   -65.47%  (p=0.000 n=10+10)
Unmarshal/[]uint64#01-8                       207ns ± 8%      60ns ±16%   -70.82%  (p=0.000 n=9+10)
Unmarshal/[]int64-8                           202ns ±20%      64ns ±10%   -68.57%  (p=0.000 n=10+10)
Unmarshal/[]int64#01-8                        197ns ±19%      56ns ± 9%   -71.50%  (p=0.000 n=10+9)
Unmarshal/[]float32-8                         197ns ±14%      61ns ±16%   -68.78%  (p=0.000 n=10+9)
Unmarshal/[]float64-8                         202ns ±16%      64ns ± 7%   -68.32%  (p=0.000 n=10+9)
Unmarshal/[]bool-8                            195ns ±15%      61ns ± 6%   -68.76%  (p=0.000 n=10+9)
Unmarshal/[]string-8                          249ns ±10%     109ns ± 5%   -56.19%  (p=0.000 n=10+10)
Unmarshal/[]amqp.symbol-8                     253ns ± 9%     110ns ± 5%   -56.60%  (p=0.000 n=10+10)
Unmarshal/[][]uint8-8                         285ns ± 3%     139ns ± 6%   -51.29%  (p=0.000 n=8+10)
Unmarshal/[]time.Time-8                       192ns ±15%      61ns ± 6%   -68.44%  (p=0.000 n=10+10)
Unmarshal/[]amqp.UUID-8                       192ns ±12%      61ns ±12%   -67.98%  (p=0.000 n=10+10)
Unmarshal/[]interface_{}-8                    296ns ± 8%     142ns ± 4%   -51.92%  (p=0.000 n=9+10)

name                                       old alloc/op   new alloc/op   delta
FrameMarshal/transfer-8                       0.00B          0.00B           ~     (all equal)
FrameUnmarshal/transfer-8                      312B ± 0%      232B ± 0%   -25.64%  (p=0.000 n=10+10)
Marshal/*amqp.performOpen-8                    115B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performBegin-8                   115B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performAttach-8                  576B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/amqp.role-8                           0.00B          0.00B           ~     (all equal)
Marshal/*amqp.unsettled-8                      115B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.source-8                         230B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.target-8                         115B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performFlow-8                    115B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performTransfer-8               0.00B          0.00B           ~     (all equal)
Marshal/*amqp.performDisposition-8            0.00B          0.00B           ~     (all equal)
Marshal/*amqp.performDetach-8                  131B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performDetach#01-8               617B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/amqp.ErrorCondition-8                 16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.Error-8                          131B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performEnd-8                     131B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.performClose-8                   131B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.Message-8                        464B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.MessageHeader-8                 4.00B ± 0%     0.00B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.MessageProperties-8             0.00B          0.00B           ~     (all equal)
Marshal/*amqp.stateReceived-8                 0.00B          0.00B           ~     (all equal)
Marshal/*amqp.stateAccepted-8                 0.00B          0.00B           ~     (all equal)
Marshal/*amqp.stateRejected-8                  131B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/*amqp.stateReleased-8                 0.00B          0.00B           ~     (all equal)
Marshal/*amqp.stateModified-8                  115B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
Marshal/amqp.lifetimePolicy-8                 0.00B          0.00B           ~     (all equal)
…