@vcabbage 2018-02-06
AppVeyor appears to be failing because it tries to put the commit message in a environment variable, but it's longer than what the system supports.